Cleanup compose.ts with DirectiveMetadata class
